Remove duplicate TestSIGReg class
